nexus_core/requirements.txt
discord.py>=2.3.0
yfinance>=0.2.31
pandas>=2.0.0
pandas-ta>=0.3.14
py_vollib>=1.0.1
pandas_market_calendars>=4.3.1
python-dotenv>=1.0.0
tzdata>=2023.3
numpy>=1.24.0
pydantic>=2.12.5
openai>=2.21.0
httpx